+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add users.can
 - add uuid to browse table
 - add name parameter to store_of
+- trap NameError when logging database close at shutdown
 
 ## [6.14.0] - 2020-12-21
 - add api module